+++ .travis.yml
@@ -4,4 +4,5 @@ jdk:
 branches:
   only:
     - master
-  
+services:
+  - redis-server
