+++ pom.xml
@@ -28,7 +28,6 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<java.version>1.7</java.version>
-		<skipTests>true</skipTests>
 	</properties>
 
 	<repositories>
@@ -103,7 +102,6 @@
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>2.13</version>
 				<configuration>
-					<skipTests>${skipTests}</skipTests>
 					<excludes>
 						<exclude>**/api/**</exclude>
 						<exclude>**/load/**</exclude>
