+++ src/test/java/oncue/AgentPresenceTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional;
+package oncue;
 
 import java.util.Arrays;
