+++ src/test/java/oncue/AgentShutdownTest.java
@@ -13,45 +13,73 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional;
-
-import static junit.framework.Assert.assertEquals;
+package oncue;
 
 import java.util.Arrays;
 
 import oncue.agent.UnlimitedCapacityAgent;
-import oncue.base.AbstractActorSystemTest;
-import oncue.messages.internal.AbstractWorkRequest;
 import oncue.messages.internal.SimpleMessages.SimpleMessage;
-import oncue.messages.internal.WorkResponse;
 import oncue.scheduler.SimpleQueuePopScheduler;
+import oncue.settings.Settings;
+import oncue.settings.SettingsProvider;
 import oncue.workers.TestWorker;
 
 import org.junit.Test;
 
 import sun.management.Agent;
 import akka.actor.Actor;
+import akka.actor.ActorSystem;
 import akka.actor.Props;
 import akka.actor.UntypedActorFactory;
+import akka.event.Logging;
+import akka.event.LoggingAdapter;
 import akka.testkit.JavaTestKit;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
+
 /**
- * Agents should register themselves with the central Scheduler when they start
- * up. After receiving confirmation that they are registered, they should
- * request work.
+ * When a remote agent shuts down gracefully (i.e. a final remote client
+ * shutdown event is broadcast), the scheduler should note this and deregister
+ * the agent, before it becomes a dead agent.
  */
-public class AgentRegistrationTest extends AbstractActorSystemTest {
+public class AgentShutdownTest {
+
+	static Config config;
+	final ActorSystem system;
+	final ActorSystem agentSystem;
+	final Settings settings;
+	LoggingAdapter log;
+
+	public AgentShutdownTest() {
+		config = ConfigFactory.load("agent-shutdown-test");
+		system = ActorSystem.create("oncue-service", config.getConfig("service").withFallback(config));
+		agentSystem = ActorSystem.create("oncue-agent", config.getConfig("client").withFallback(config));
+		settings = SettingsProvider.SettingsProvider.get(system);
+		log = Logging.getLogger(system, this);
+	}
 
 	/**
 	 * An {@linkplain Agent} should emit a steady heartbeat while it is alive.
 	 */
 	@SuppressWarnings("serial")
 	@Test
-	public void agentRegistersAndRequestsWorkButReceivesNoWork() {
+	public void agentShutsDownGracefully() {
 		new JavaTestKit(system) {
 			{
 				// Create a simple scheduler with a probe
-				final JavaTestKit schedulerProbe = new JavaTestKit(system);
+				final JavaTestKit schedulerProbe = new JavaTestKit(system) {
+					{
+						new IgnoreMsg() {
+
+							@Override
+							protected boolean ignore(Object message) {
+								return !(message.equals(SimpleMessage.AGENT_SHUTDOWN));
+							}
+						};
+					}
+				};
+
 				system.actorOf(new Props(new UntypedActorFactory() {
 					@Override
 					public Actor create() throws Exception {
@@ -63,7 +91,7 @@ public class AgentRegistrationTest extends AbstractActorSystemTest {
 
 				// Create an agent with a probe
 				final JavaTestKit agentProbe = new JavaTestKit(system);
-				system.actorOf(new Props(new UntypedActorFactory() {
+				agentSystem.actorOf(new Props(new UntypedActorFactory() {
 					@Override
 					public Actor create() throws Exception {
 						UnlimitedCapacityAgent agent = new UnlimitedCapacityAgent(Arrays.asList(TestWorker.class
@@ -73,18 +101,14 @@ public class AgentRegistrationTest extends AbstractActorSystemTest {
 					}
 				}), "agent");
 
-				// Expect the initial heartbeat from the agent
-				schedulerProbe.expectMsgEquals(SimpleMessage.AGENT_HEARTBEAT);
-
-				// Expect the registration message from the scheduler
+				// Wait until the agent is registered
 				agentProbe.expectMsgEquals(SimpleMessage.AGENT_REGISTERED);
 
-				// Expect the agent to request work
-				schedulerProbe.expectMsgClass(AbstractWorkRequest.class);
+				// Tell the agent to stop
+				agentSystem.shutdown();
 
-				// Expect no work from the scheduler
-				WorkResponse workResponse = agentProbe.expectMsgClass(WorkResponse.class);
-				assertEquals("Expected no jobs", 0, workResponse.getJobs().size());
+				// Expect the agent shutdown message
+				schedulerProbe.expectMsgEquals(duration("5 seconds"), SimpleMessage.AGENT_SHUTDOWN);
 			}
 		};
 	}
