+++ src/test/java/oncue/EnqueueJobTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional;
+package oncue;
 
 import static org.junit.Assert.assertEquals;
 import oncue.base.AbstractActorSystemTest;
