+++ src/test/java/oncue/JobProgressTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional.robustness;
+package oncue;
 
 import java.util.Arrays;
 
@@ -21,47 +21,49 @@ import junit.framework.Assert;
 import oncue.agent.UnlimitedCapacityAgent;
 import oncue.base.AbstractActorSystemTest;
 import oncue.messages.internal.EnqueueJob;
-import oncue.messages.internal.Job;
 import oncue.messages.internal.JobProgress;
-import oncue.messages.internal.SimpleMessages.SimpleMessage;
 import oncue.queueManager.InMemoryQueueManager;
 import oncue.scheduler.SimpleQueuePopScheduler;
 import oncue.workers.TestWorker;
 
 import org.junit.Test;
 
-import sun.management.Agent;
 import akka.actor.Actor;
 import akka.actor.ActorRef;
-import akka.actor.PoisonPill;
 import akka.actor.Props;
 import akka.actor.UntypedActorFactory;
 import akka.testkit.JavaTestKit;
 
 /**
- * An {@linkplain Agent} may die unexpectedly while it is processing jobs. Since
- * we cannot rely on the agent sending a message in its death-throes (the entire
- * JVM may have exploded), we need to detect that its heart beat has stopped and
- * take action back at the scheduler.
+ * When a job has been scheduled, the worker may elect to send back regular
+ * reports about job completion progress.
  */
-public class AgentDiesTest extends AbstractActorSystemTest {
+public class JobProgressTest extends AbstractActorSystemTest {
 
 	@Test
 	@SuppressWarnings("serial")
-	public void testAgentDiesAndAnotherReplacesIt() {
+	public void monitorProgress() {
 		new JavaTestKit(system) {
 			{
 				// Create a scheduler probe
 				final JavaTestKit schedulerProbe = new JavaTestKit(system) {
 					{
 						new IgnoreMsg() {
-
 							@Override
 							protected boolean ignore(Object message) {
-								if (message.equals(SimpleMessage.AGENT_DEAD) || message instanceof JobProgress)
-									return false;
-								else
-									return true;
+								return !(message instanceof JobProgress);
+							}
+						};
+					}
+				};
+
+				// Create an agent probe
+				final JavaTestKit agentProbe = new JavaTestKit(system) {
+					{
+						new IgnoreMsg() {
+							@Override
+							protected boolean ignore(Object message) {
+								return !(message instanceof JobProgress);
 							}
 						};
 					}
@@ -71,7 +73,7 @@ public class AgentDiesTest extends AbstractActorSystemTest {
 				ActorRef queueManager = system.actorOf(new Props(InMemoryQueueManager.class),
 						settings.QUEUE_MANAGER_NAME);
 
-				// Create a simple scheduler
+				// Create a scheduler with a probe
 				system.actorOf(new Props(new UntypedActorFactory() {
 					@Override
 					public Actor create() throws Exception {
@@ -81,42 +83,32 @@ public class AgentDiesTest extends AbstractActorSystemTest {
 					}
 				}), settings.SCHEDULER_NAME);
 
-				// Create an agent
-				ActorRef agent1 = system.actorOf(new Props(new UntypedActorFactory() {
-					@Override
-					public Actor create() throws Exception {
-						return new UnlimitedCapacityAgent(Arrays.asList(TestWorker.class.getName()));
-					}
-				}), "agent1");
-
 				// Enqueue a job
 				queueManager.tell(new EnqueueJob(TestWorker.class.getName()), getRef());
-				Job job = expectMsgClass(Job.class);
 
-				// Wait for some progress
-				schedulerProbe.expectMsgClass(JobProgress.class);
-
-				// Tell the agent to commit seppuku
-				agent1.tell(PoisonPill.getInstance(), getRef());
-
-				// Expect a message about agent death
-				schedulerProbe.expectMsgEquals(settings.SCHEDULER_AGENT_HEARTBEAT_TIMEOUT.plus(duration("5 seconds")),
-						SimpleMessage.AGENT_DEAD);
-
-				// The heartbeat of the original agent should die
-				schedulerProbe.expectNoMsg(settings.AGENT_HEARTBEAT_FREQUENCY);
-
-				// Now, create a second agent
+				// Start an agent with a probe
 				system.actorOf(new Props(new UntypedActorFactory() {
 					@Override
 					public Actor create() throws Exception {
-						return new UnlimitedCapacityAgent(Arrays.asList(TestWorker.class.getName()));
+						UnlimitedCapacityAgent agent = new UnlimitedCapacityAgent(Arrays.asList(TestWorker.class
+								.getName()));
+						agent.injectProbe(agentProbe.getRef());
+						return agent;
 					}
-				}), "agent2");
+				}), settings.AGENT_NAME);
 
-				// Wait for some progress on the original job
-				JobProgress jobProgress = schedulerProbe.expectMsgClass(JobProgress.class);
-				Assert.assertEquals("Wrong job being worked on", job.getId(), jobProgress.getJob().getId());
+				// Expect a series of progress reports
+				double expectedProgress = 0;
+				for (int i = 0; i < 5; i++) {
+					JobProgress agentProgress = agentProbe.expectMsgClass(JobProgress.class);
+					JobProgress schedulerProgress = schedulerProbe.expectMsgClass(JobProgress.class);
+					Assert.assertEquals(expectedProgress, agentProgress.getProgress());
+					Assert.assertEquals(expectedProgress, schedulerProgress.getProgress());
+					expectedProgress += 0.25;
+				}
+
+				// Expect no more progress
+				schedulerProbe.expectNoMsg();
 			}
 		};
 	}
