+++ src/test/java/oncue/WorkerTest.java
@@ -1,26 +1,23 @@
 /*******************************************************************************
  * Copyright 2013 Michael Marconi
  * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  * 
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  ******************************************************************************/
-package oncue.functional;
+package oncue;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertFalse;
 
 import java.util.Arrays;
 
-import junit.framework.Assert;
 import oncue.agent.UnlimitedCapacityAgent;
 import oncue.base.AbstractActorSystemTest;
 import oncue.messages.internal.EnqueueJob;
@@ -42,9 +39,8 @@ import akka.testkit.JavaTestKit;
 import akka.testkit.TestActorRef;
 
 /**
- * When an {@linkplain Agent} receives a {@linkplain WorkResponse}, it will
- * attempt to spawn an instance of an {@linkplain IWorker} for each
- * {@linkplain Job} in the list.
+ * When an {@linkplain Agent} receives a {@linkplain WorkResponse}, it will attempt to spawn an
+ * instance of an {@linkplain IWorker} for each {@linkplain Job} in the list.
  */
 public class WorkerTest extends AbstractActorSystemTest {
 
@@ -52,16 +48,20 @@ public class WorkerTest extends AbstractActorSystemTest {
 	@SuppressWarnings("serial")
 	public void spawnWorkerAndStartJob() {
 		new JavaTestKit(system) {
+
 			{
 				// Create an agent probe
 				final JavaTestKit agentProbe = new JavaTestKit(system) {
+
 					{
 						new IgnoreMsg() {
+
 							protected boolean ignore(Object message) {
-								if (message instanceof WorkResponse || message instanceof JobProgress)
+								if (message instanceof WorkResponse
+										|| message instanceof JobProgress)
 									return false;
-								else
-									return true;
+
+								return true;
 							}
 						};
 					}
@@ -81,8 +81,9 @@ public class WorkerTest extends AbstractActorSystemTest {
 				}), settings.SCHEDULER_NAME);
 
 				// Create and expose an agent
-				TestActorRef<UnlimitedCapacityAgent> agentRef = TestActorRef.create(system, new Props(
-						new UntypedActorFactory() {
+				TestActorRef<UnlimitedCapacityAgent> agentRef = TestActorRef.create(system,
+						new Props(new UntypedActorFactory() {
+
 							@Override
 							public Actor create() throws Exception {
 								UnlimitedCapacityAgent agent = new UnlimitedCapacityAgent(Arrays
@@ -95,10 +96,11 @@ public class WorkerTest extends AbstractActorSystemTest {
 
 				// Wait until the agent receives an empty work response
 				WorkResponse workResponse = agentProbe.expectMsgClass(WorkResponse.class);
-				Assert.assertEquals(0, workResponse.getJobs().size());
+				assertEquals(0, workResponse.getJobs().size());
 
 				// Check that there are no workers
-				assertFalse("Expected no child workers", agent.getContext().getChildren().iterator().hasNext());
+				assertFalse("Expected no child workers", agent.getContext().getChildren()
+						.iterator().hasNext());
 
 				// Enqueue a job
 				queueManager.tell(new EnqueueJob(TestWorker.class.getName()), getRef());
@@ -114,6 +116,7 @@ public class WorkerTest extends AbstractActorSystemTest {
 
 				// Wait for the job to complete
 				new AwaitCond(duration("5 seconds")) {
+
 					@Override
 					protected boolean cond() {
 						return agentProbe.expectMsgClass(JobProgress.class).getProgress() == 1.0;
@@ -122,6 +125,7 @@ public class WorkerTest extends AbstractActorSystemTest {
 
 				// Ensure the worker is dead
 				new AwaitCond(duration("5 seconds")) {
+
 					@Override
 					protected boolean cond() {
 						return worker.isTerminated() == true;
