+++ src/test/java/oncue/redis/RedisQueueManagerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional.redis;
+package oncue.redis;
 
 import static junit.framework.Assert.assertEquals;
 import oncue.backingstore.RedisBackingStore;
