+++ src/test/java/oncue/robustness/RobustRedisDequeueTest.java
@@ -1,4 +1,4 @@
-package oncue.functional.robustness;
+package oncue.robustness;
 
 import static junit.framework.Assert.assertEquals;
 import oncue.backingstore.RedisBackingStore;
@@ -11,6 +11,7 @@ import oncue.workers.TestWorker;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import redis.clients.jedis.Jedis;
@@ -33,6 +34,7 @@ public class RobustRedisDequeueTest extends AbstractActorSystemTest {
 	private Jedis redis;
 
 	@Test
+	@Ignore
 	@SuppressWarnings("serial")
 	public void atomicPopAndPush() {
 		new JavaTestKit(system) {
@@ -77,6 +79,7 @@ public class RobustRedisDequeueTest extends AbstractActorSystemTest {
 	}
 
 	@Test
+	@Ignore
 	@SuppressWarnings("serial")
 	public void bypassAddingUnscheduledJob() {
 		new JavaTestKit(system) {
