+++ src/test/java/oncue/robustness/WorkerDiesTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package oncue.functional.robustness;
+package oncue.robustness;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertTrue;
