+++ src/test/java/oncue/strategies/JVMCapacityStrategyTest.java
@@ -1,19 +1,17 @@
 /*******************************************************************************
  * Copyright 2013 Michael Marconi
  * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  * 
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  ******************************************************************************/
-package oncue.functional.strategies;
+package oncue.strategies;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -38,40 +36,41 @@ import akka.testkit.JavaTestKit;
 import akka.testkit.TestActorRef;
 
 /**
- * Test the JVM memory capacity strategy by farming jobs of known size out to
- * agents with known capacities.
+ * Test the JVM memory capacity strategy by farming jobs of known size out to agents with known
+ * capacities.
  */
 public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 
 	/*-
 	 *            Jobs: J1(100), J2(200), J3(50), J4(200), J5(500) 
 	 *  Agent capacity: A1(500), A2(70), A3(200)
-	 *  
+	 *
 	 *  J1(100): | A2(70) | A3(200) | A1(500) |
 	 *           | x      | J1      | -       |
-	 *           
+	 *
 	 *  J2(200): | A2(70) | A3(100) | A1(500) |
 	 *           | x      | x       | J2      |
-	 *           
+	 *
 	 *  J3(50):  | A2(70) | A3(100) | A1(300) |
 	 *           | J3     | -       | -       |
-	 *           
+	 *
 	 *  J4(200): | A2(20) | A3(100) | A1(300) |
 	 *           | x      | x       | J4      |
-	 *           
+	 *
 	 *  J5(500): | A2(20) | A3(100) | A1(100) |
 	 *           | x      | x       | x       |
-	 *           
+	 *
 	 *  Final allocation:  A1(J2,J4)
 	 *                     A2(J3)
 	 *                     A3(J1)
-	 *                     Job 5 unscheduled!                                                      
+	 *                     Job 5 unscheduled!
 	 */
 
 	@Test
 	@SuppressWarnings("serial")
 	public void jvmCapacityStrategyTest() {
 		new JavaTestKit(system) {
+
 			{
 				// Create an in-memory queue manager
 				ActorRef queueManager = system.actorOf(new Props(InMemoryQueueManager.class),
@@ -79,12 +78,14 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 
 				// Create a scheduler probe
 				final JavaTestKit schedulerProbe = new JavaTestKit(system) {
+
 					{
 						new IgnoreMsg() {
 
 							@Override
 							protected boolean ignore(Object message) {
 								return false;
+								// TODO Is this still needed?
 								// return !(message instanceof JobProgress ||
 								// message instanceof Job);
 							}
@@ -94,6 +95,7 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 
 				// Create a naked JVM capacity-aware scheduler
 				final Props schedulerProps = new Props(new UntypedActorFactory() {
+
 					@Override
 					public Actor create() throws Exception {
 						JVMCapacityScheduler scheduler = new JVMCapacityScheduler(null);
@@ -101,8 +103,8 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 						return scheduler;
 					}
 				});
-				final TestActorRef<JVMCapacityScheduler> schedulerRef = TestActorRef.create(system, schedulerProps,
-						settings.SCHEDULER_NAME);
+				final TestActorRef<JVMCapacityScheduler> schedulerRef = TestActorRef.create(system,
+						schedulerProps, settings.SCHEDULER_NAME);
 				final JVMCapacityScheduler scheduler = schedulerRef.underlyingActor();
 
 				// Create agent probes
@@ -111,8 +113,7 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 				final JavaTestKit agentProbe3 = createAgentProbe();
 
 				/*
-				 * Create three capacity-aware agents with pre-determined
-				 * capacity
+				 * Create three capacity-aware agents with pre-determined capacity
 				 */
 				createAgent(agentProbe1, "agent1", 500);
 				createAgent(agentProbe2, "agent2", 70);
@@ -124,8 +125,8 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 				agentProbe3.expectMsgClass(WorkResponse.class);
 
 				/*
-				 * Pause the scheduler, as we want the scheduler to see all
-				 * enqueued jobs in a single batch for testing purposes
+				 * Pause the scheduler, as we want the scheduler to see all enqueued jobs in a
+				 * single batch for testing purposes
 				 */
 				scheduler.pause();
 
@@ -155,43 +156,44 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 				Assert.assertEquals(2, agent1WorkResponse.getJobs().size());
 
 				Job agent1Job2 = agent1WorkResponse.getJobs().get(0);
-				Assert.assertEquals(job2.getParams().get(JVMCapacityScheduler.JOB_SIZE),
-						agent1Job2.getParams().get(JVMCapacityScheduler.JOB_SIZE));
+				Assert.assertEquals(job2.getParams().get(JVMCapacityScheduler.JOB_SIZE), agent1Job2
+						.getParams().get(JVMCapacityScheduler.JOB_SIZE));
 
 				Job agent1Job4 = agent1WorkResponse.getJobs().get(0);
-				Assert.assertEquals(job4.getParams().get(JVMCapacityScheduler.JOB_SIZE),
-						agent1Job4.getParams().get(JVMCapacityScheduler.JOB_SIZE));
+				Assert.assertEquals(job4.getParams().get(JVMCapacityScheduler.JOB_SIZE), agent1Job4
+						.getParams().get(JVMCapacityScheduler.JOB_SIZE));
 
 				// Expect Job 3 at Agent 2
 				WorkResponse agent2WorkResponse = agentProbe2.expectMsgClass(WorkResponse.class);
 				Assert.assertEquals(1, agent2WorkResponse.getJobs().size());
 				Job agent2Job3 = agent2WorkResponse.getJobs().get(0);
-				Assert.assertEquals(job3.getParams().get(JVMCapacityScheduler.JOB_SIZE),
-						agent2Job3.getParams().get(JVMCapacityScheduler.JOB_SIZE));
+				Assert.assertEquals(job3.getParams().get(JVMCapacityScheduler.JOB_SIZE), agent2Job3
+						.getParams().get(JVMCapacityScheduler.JOB_SIZE));
 
 				// Expect Job 1 at Agent 3
 				WorkResponse agent3WorkResponse = agentProbe3.expectMsgClass(WorkResponse.class);
 				Assert.assertEquals(1, agent3WorkResponse.getJobs().size());
 				Job agent3Job1 = agent3WorkResponse.getJobs().get(0);
-				Assert.assertEquals(job1.getParams().get(JVMCapacityScheduler.JOB_SIZE),
-						agent3Job1.getParams().get(JVMCapacityScheduler.JOB_SIZE));
+				Assert.assertEquals(job1.getParams().get(JVMCapacityScheduler.JOB_SIZE), agent3Job1
+						.getParams().get(JVMCapacityScheduler.JOB_SIZE));
 
 				/*
-				 * Now, Jobs 1 - 4 will complete and all the agents will ask for
-				 * new work. Since Agent 1 is the only one big enough to fit Job
-				 * 5, it will receive it.
+				 * Now, Jobs 1 - 4 will complete and all the agents will ask for new work. Since
+				 * Agent 1 is the only one big enough to fit Job 5, it will receive it.
 				 */
-				agent1WorkResponse = agentProbe1.expectMsgClass(duration("5 seconds"), WorkResponse.class);
+				agent1WorkResponse = agentProbe1.expectMsgClass(duration("5 seconds"),
+						WorkResponse.class);
 				Assert.assertEquals(1, agent1WorkResponse.getJobs().size());
 				Job agent1Job5 = agent1WorkResponse.getJobs().get(0);
-				Assert.assertEquals(job5.getParams().get(JVMCapacityScheduler.JOB_SIZE),
-						agent1Job5.getParams().get(JVMCapacityScheduler.JOB_SIZE));
+				Assert.assertEquals(job5.getParams().get(JVMCapacityScheduler.JOB_SIZE), agent1Job5
+						.getParams().get(JVMCapacityScheduler.JOB_SIZE));
 			}
 		};
 	}
 
 	private JavaTestKit createAgentProbe() {
 		return new JavaTestKit(system) {
+
 			{
 				new IgnoreMsg() {
 
@@ -207,9 +209,11 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 	@SuppressWarnings("serial")
 	private void createAgent(final JavaTestKit agentProbe, String name, final int capacity) {
 		system.actorOf(new Props(new UntypedActorFactory() {
+
 			@Override
 			public Actor create() throws Exception {
-				JVMCapacityAgent agent = new JVMCapacityAgent(Arrays.asList(TestWorker.class.getName()), capacity);
+				JVMCapacityAgent agent = new JVMCapacityAgent(Arrays.asList(TestWorker.class
+						.getName()), capacity);
 				agent.injectProbe(agentProbe.getRef());
 				return agent;
 			}
@@ -219,6 +223,7 @@ public class JVMCapacityStrategyTest extends AbstractActorSystemTest {
 	@SuppressWarnings("serial")
 	private void enqueueJob(ActorRef queueManager, ActorRef testKit, final int jobSize) {
 		queueManager.tell(new EnqueueJob(TestWorker.class.getName(), new HashMap<String, String>() {
+
 			{
 				put(JVMCapacityScheduler.JOB_SIZE, new Integer(jobSize).toString());
 			}
